`ast.hpp`: Finish `UnOp` semantic analysis
We completed the `sem()` function of class `UnOp` for cases: "head",
"tail" and "nil"?. Additionally, we added `TYPE_any` in `TypeBlock`.
It will help us describe the `nil` constant (empty list of any type
of any type of elements).

This commit also contains example programs to test the new functionality.