ast: Add function `TheMalloc`
We will use this function to allocate memory using LLVM. We need
this in order allocate memory blocks that are more complex than
simple ints or chars.

For example, a list element has (1) a value and (2) a pointer to
the next element in the list. We will use 'TheMalloc' to allocate
memory for such structures.

Note: Nickie uses 'TheMalloc' for minibasic too.